add collarCommandFactory; move rampsensor into collar folder
